refactor: extract section components from AppSettingsMenu
Split the 520-line VisualEffectsMenu/index.tsx into focused modules:
- SourcesSections.tsx: MusicSourcesSection and NativeQueueSyncSection
- ProviderDataSection.tsx: per-provider data management UI
- CollapsibleSection.tsx: reusable collapsible wrapper

Closes #421
